requirements.txt
alembic==1.6.5
anyio==3.2.1
apache-airflow==2.1.2
apache-airflow-providers-amazon==2.1.0
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-google==5.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-mysql==2.1.0
apache-airflow-providers-sqlite==2.0.0
apiclient==1.0.4
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
blinker==1.4
boa-str==1.1.0
boto3==1.17.112
botocore==1.20.112
cachetools==4.2.2
cattrs==1.5.0
certifi==2020.12.5
cffi==1.14.5
charset-normalizer==2.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.7
defusedxml==0.7.1
dill==0.3.1.1
distlib==0.3.2
dnspython==1.16.0
docutils==0.16
email-validator==1.1.3
filelock==3.0.12
Flask==1.1.4
Flask-AppBuilder==3.3.1
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
google-ads==14.0.0
google-api-core==1.31.2
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-cloud-appengine-logging==0.1.4
google-cloud-audit-log==0.1.1
google-cloud-automl==2.4.2
google-cloud-bigquery==2.25.1
google-cloud-bigquery-datatransfer==3.3.2
google-cloud-bigquery-storage==2.6.3
google-cloud-bigtable==1.7.0
google-cloud-container==1.0.1
google-cloud-core==1.7.2
google-cloud-datacatalog==3.4.0
google-cloud-dataproc==2.5.0
google-cloud-dlp==1.0.0
google-cloud-kms==2.6.0
google-cloud-language==1.3.0
google-cloud-logging==2.6.0
google-cloud-memcache==1.0.0
google-cloud-monitoring==2.5.0
google-cloud-os-login==2.3.1
google-cloud-pubsub==2.7.1
google-cloud-redis==2.2.2
google-cloud-secret-manager==1.0.0
google-cloud-spanner==1.19.1
google-cloud-speech==1.3.2
google-cloud-storage==1.42.0
google-cloud-tasks==2.5.1
google-cloud-texttospeech==1.0.1
google-cloud-translate==1.7.0
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-cloud-workflows==1.2.1
google-crc32c==1.1.3
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
graphviz==0.16
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
grpcio-gcp==0.2.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httplib2==0.19.1
httpx==0.18.2
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
libcst==0.3.20
lockfile==0.12.2
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.12.2
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.0
marshmallow-sqlalchemy==0.23.1
multitasking==0.0.9
mypy-extensions==0.4.3
mysql-connector-python==8.0.22
mysqlclient==2.0.3
nox==2020.12.31
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.0
pandas==1.3.0
pandas-gbq==0.14.1
pendulum==2.1.2
platformdirs==2.3.0
prison==0.1.3
proto-plus==1.19.0
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.2.0
Pygments==2.9.0
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.5.0
rsa==4.7.2
s3transfer==0.4.2
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
typing-extensions==3.10.0.2
typing-inspect==0.7.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
watchtower==1.0.6
Werkzeug==1.0.1
WTForms==2.3.3
yfinance==0.1.63
zipp==3.5.0
